Cambios menores
Cambios varios.

+++ assets/js/foods.js
@@ -746,6 +746,7 @@ var seleccioninicial = $("#dynamicnumpersonselect option:selected").val();
 //var inginicial=$("#espan").text();
 
 function dynamicnumperson(){
+
 	$("ul.list-group li").each(function( index ){
 
 	 var seleccion = $("#dynamicnumpersonselect option:selected").val();
@@ -753,20 +754,20 @@ function dynamicnumperson(){
 	 //console.log(ing)
 
 	 if (seleccion>parseInt(seleccioninicial)) {
-	 	var x = ((parseInt(seleccion)*parseInt(ing))/parseInt(seleccioninicial)).toFixed(2).toString().split(".");
+	 	var x = ((parseInt(seleccion)*parseFloat(ing))/parseInt(seleccioninicial)).toFixed(2).toString().split(".");
 	 	if (x[1]=="00") {
-	 		$(this).find("span").text(((parseInt(seleccion)*parseInt(ing))/parseInt(seleccioninicial)).toFixed(0));
+	 		$(this).find("span").text(((parseInt(seleccion)*parseFloat(ing))/parseInt(seleccioninicial)).toFixed(0));
 	 	}else{
-	 		$(this).find("span").text(((parseInt(seleccion)*parseInt(ing))/parseInt(seleccioninicial)).toFixed(2));
+	 		$(this).find("span").text(((parseInt(seleccion)*parseFloat(ing))/parseInt(seleccioninicial)).toFixed(2));
 	 	}
 
 
 	 }else if(seleccion<parseInt(seleccioninicial)){
-	 	var y = ((parseInt(seleccion)*parseInt(ing))/parseInt(seleccioninicial)).toFixed(2).toString().split(".");
+	 	var y = ((parseInt(seleccion)*parseFloat(ing))/parseInt(seleccioninicial)).toFixed(2).toString().split(".");
 	 	if (y[1]=="00") {
-	 		$(this).find("span").text(((parseInt(seleccion)*parseInt(ing))/parseInt(seleccioninicial)).toFixed(0));
+	 		$(this).find("span").text(((parseInt(seleccion)*parseFloat(ing))/parseInt(seleccioninicial)).toFixed(0));
 	 	}else{
-	 		$(this).find("span").text(((parseInt(seleccion)*parseInt(ing))/parseInt(seleccioninicial)).toFixed(2));
+	 		$(this).find("span").text(((parseInt(seleccion)*parseFloat(ing))/parseInt(seleccioninicial)).toFixed(2));
 	 	}
